+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.29.2
 
+* `Core`:
+    * New dice type `SlotMachineDiceAnimationType`
+
 ## 0.29.1
 
 * `Common`: