+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `Common`:
     * Version updates:
         * `Coroutines`: `1.3.9` -> `1.4.0`
+    * Internal broadcast channels were replaced with `SharedFlow`
+* `TelegramBotAPI-extensions-utils`:
+    * Extension `ReceiveChannel#debounceByValue` has been deprecated
 
 ## 0.29.1